--- App.tsx
@@ -1,9 +1,0 @@
-import React from 'react';
-import LoginScreen from './screens/login';
-import CadastrarServico from './screens/servicos';
-
-const App = () => {
-  return <CadastrarServico />;
-};
-
-export default App;
