+++ app/pecas/editar.tsx
@@ -1,20 +1,25 @@
 import React, { useState } from 'react';
 import { View, Text,} from 'react-native';
-import { globalStyles } from '../../styles/globalStyles';
+import { useRouter } from 'expo-router';
+import { Image } from 'react-native';
+
 import { CustomInput } from '../../components/CustomInput';
 import { ExpandingTextArea } from '../../components/ExpandingTextArea';
-import CustomButton from '../../components/CustomButton';
-import ImagePickerInput from '../../components/ImagePickerInput';
+import { CustomButton } from '../../components/CustomButton';
+import { ImagePickerInput } from '../../components/ImagePickerInput';
+
+import { globalStyles } from '../../styles/globalStyles';
 import { pecStyles } from './styles';
-import { Image } from 'react-native';
 import { formatarPreco } from '../../utils/formatters';
 
-const CadastrarPecas = ({ navigation }: any) => {
-    const [nome, setNome] = useState('');
-    const [descricao, setDescricao] = useState('');
-    const [quantidade, setQuantidade] = useState('');
-    const [fabricante, setFabricante] = useState('');
-    const [preco, setPreco] = useState('');
+const EditarPecas = () => {
+    const router = useRouter();
+
+    const [nome, setNome] = useState('Pneu');
+    const [descricao, setDescricao] = useState('Componente essencial para a segurança e desempenho do veículo, o pneu é responsável pelo contato direto com o solo, garantindo aderência, estabilidade e conforto na condução. Recomendado verificar periodicamente seu estado e calibragem para evitar desgastes irregulares e riscos de acidentes.');
+    const [quantidade, setQuantidade] = useState('16');
+    const [fabricante, setFabricante] = useState('Michelin');
+    const [preco, setPreco] = useState('500');
 
 
     const [imagem, setImagem] = useState<string | null>(null);
@@ -24,15 +29,15 @@ const CadastrarPecas = ({ navigation }: any) => {
       setPreco(precoFormatado);
     };
 
-    const handleCadastrar = () => {
+    const handleEditar = () => {
         console.log({
-        nome,
-        descricao,
-        fabricante,
-        quantidade,
-        preco,
+            nome,
+            descricao,
+            fabricante,
+            quantidade,
+            preco,
         });
-        navigation.goBack();
+        router.back(); 
     };
 
 
@@ -44,10 +49,11 @@ const CadastrarPecas = ({ navigation }: any) => {
         </View>
               
         <View style={pecStyles.initialBottom}>
-            <Text style={globalStyles.title}>Cadastrar Peça</Text>
+            <Text style={globalStyles.title}>Editar Peça</Text>
             <CustomInput
                 label="Nome"
-                placeholder="Digite o nome do peciço"
+                value={nome}
+                placeholder="Digite o nome da peça"
                 onChangeText={setNome}
                 contentStyle={{ width: '80%', maxWidth: 400 }}
             />
@@ -93,13 +99,17 @@ const CadastrarPecas = ({ navigation }: any) => {
 
             <View style={pecStyles.crudButtons}>
                 <CustomButton 
-                    style={{width: '39%', maxWidth: 193, height: 50}} 
-                    title="Cadastrar" 
-                    onPress={handleCadastrar} />
+                    style={{width: '25%', maxWidth: 127, height: 50}} 
+                    title="Salvar" 
+                    onPress={() => router.back()} />
                 <CustomButton 
-                    style={{width: '39%', maxWidth: 193, height: 50}} 
+                    style={{width: '25%', maxWidth: 127, height: 50}} 
+                    title="Deletar" 
+                    onPress={() => router.back()} />
+                <CustomButton 
+                    style={{width: '25%', maxWidth: 127, height: 50}} 
                     title="Cancelar" 
-                    onPress={handleCadastrar} />
+                    onPress={() => router.back()} />
             </View>
 
         </View>
@@ -108,4 +118,4 @@ const CadastrarPecas = ({ navigation }: any) => {
 };
 
 
-export default CadastrarPecas;
+export default EditarPecas;
