+++ app/pecas/styles.tsx
