+++ app/pecas/visualizar.tsx
@@ -1,16 +1,22 @@
 import React, { useState } from 'react';
 import { View, Text, Image } from 'react-native';
+import { useRouter } from 'expo-router';
+
+import { InfoView } from '../../components/InfoView';
+import { CustomButton } from '../../components/CustomButton';
+
 import { globalStyles } from '../../styles/globalStyles';
 import { pecStyles } from './styles';
-import InfoView from '../../components/InfoView';
-import CustomButton from '../../components/CustomButton';
 
-const VisualizarPeca = ({ navigation }: any) => {
-  const [nome] = useState('Pneu Michelin');
-  const [descricao] = useState('Componente essencial para a segurança e desempenho do veículo, o pneu é responsável pelo contato direto com o solo, garantindo aderência, estabilidade e conforto na condução. Recomendado verificar periodicamente seu estado e calibragem para evitar desgastes irregulares e riscos de acidentes.');
-  const [quantidade] = useState('16');
-  const [fabricante] = useState('Michelin');
-  const [preco] = useState('R$ 500,00');
+
+const VisualizarPeca = () => {
+    const router = useRouter();
+
+    const [nome] = useState('Pneu Michelin');
+    const [descricao] = useState('Componente essencial para a segurança e desempenho do veículo, o pneu é responsável pelo contato direto com o solo, garantindo aderência, estabilidade e conforto na condução. Recomendado verificar periodicamente seu estado e calibragem para evitar desgastes irregulares e riscos de acidentes.');
+    const [quantidade] = useState('16');
+    const [fabricante] = useState('Michelin');
+    const [preco] = useState('R$ 500,00');
 
   const handleVisualizar = () => {
     console.log({
@@ -20,6 +26,7 @@ const VisualizarPeca = ({ navigation }: any) => {
       quantidade,
       preco,
     });
+    router.back();
   };
 
   return (
@@ -45,7 +52,7 @@ const VisualizarPeca = ({ navigation }: any) => {
         <CustomButton
           style={{ width: '80%', maxWidth: 400, height: 50, marginTop: 20 }}
           title="Editar"
-          onPress={handleVisualizar}
+          onPress={() => router.push('/pecas/editar')}
         />
       </View>
     </View>
