+++ app/servicos/cadastrar.tsx
@@ -1,22 +1,27 @@
 import React, { useState } from 'react';
-import { View, Text, } from 'react-native';
-import { globalStyles } from '../../styles/globalStyles';
+import { View, Text, Image } from 'react-native';
+import { useRouter } from 'expo-router';
+import { Picker } from '@react-native-picker/picker';
+import { Checkbox } from 'expo-checkbox';
+
 import { CustomInput } from '../../components/CustomInput';
 import { ExpandingTextArea } from '../../components/ExpandingTextArea';
-import CustomButton from '../../components/CustomButton';
-import { servStyles } from './styles';
-import { Picker } from '@react-native-picker/picker';
-import Checkbox from 'expo-checkbox';
-import { formatarPreco } from '../../utils/formatters';
-import { Image } from 'react-native';
+import { CustomButton } from '../../components/CustomButton';
 
-const EditarServico = ({ navigation }: any) => {
-    const [nome, setNome] = useState('Troca de Oléo');
-    const [descricao, setDescricao] = useState('Substituição do óleo do motor por um novo, garantindo a lubrificação adequada das peças internas e o bom desempenho do veículo. Inclui verificação e, se necessário, troca do filtro de óleo. Recomendado conforme a quilometragem ou tempo de uso.');
-    const [categoria, setCategoria] = useState('mecanica');
-    const [tempoEstimado, setTempoEstimado] = useState('1');
-    const [precoMin, setPrecoMin] = useState('100');
-    const [precoMax, setPrecoMax] = useState('200');
+import { globalStyles } from '../../styles/globalStyles';
+import { servStyles } from './styles';
+import { formatarPreco } from '../../utils/formatters';
+
+
+const CadastrarServico = () => {
+    const router = useRouter();
+
+    const [nome, setNome] = useState('');
+    const [descricao, setDescricao] = useState('');
+    const [categoria, setCategoria] = useState('');
+    const [tempoEstimado, setTempoEstimado] = useState('');
+    const [precoMin, setPrecoMin] = useState('');
+    const [precoMax, setPrecoMax] = useState('');
     const [checked, setChecked] = useState(false);
 
 
@@ -30,7 +35,7 @@ const EditarServico = ({ navigation }: any) => {
       setPrecoMax(precoFormatado);
     };
 
-    const handleEditar = () => {
+    const handleCadastrar = () => {
         console.log({
         nome,
         descricao,
@@ -39,7 +44,7 @@ const EditarServico = ({ navigation }: any) => {
         precoMin,
         precoMax
         });
-        navigation.goBack();
+    router.back(); 
     };
 
 
@@ -51,11 +56,10 @@ const EditarServico = ({ navigation }: any) => {
         </View>
               
         <View style={servStyles.initialBottom}>
-            <Text style={globalStyles.title}>Editar Serviço</Text>
+            <Text style={globalStyles.title}>Cadastrar Serviço</Text>
             <CustomInput
                 label="Nome"
                 placeholder="Digite o nome do serviço"
-                value={nome}
                 onChangeText={setNome}
                 contentStyle={{ width: '80%', maxWidth: 400 }}
             />
@@ -122,17 +126,13 @@ const EditarServico = ({ navigation }: any) => {
                 </View>
             <View style={servStyles.crudButtons}>
                 <CustomButton 
-                    style={{width: '25%', maxWidth: 127, height: 50}} 
-                    title="Salvar" 
-                    onPress={handleEditar} />
+                    style={{width: '39%', maxWidth: 193, height: 50}} 
+                    title="Cadastrar" 
+                    onPress={handleCadastrar} />
                 <CustomButton 
-                    style={{width: '25%', maxWidth: 127, height: 50}} 
-                    title="Deletar" 
-                    onPress={handleEditar} />
-                <CustomButton 
-                    style={{width: '25%', maxWidth: 127, height: 50}} 
+                    style={{width: '39%', maxWidth: 193, height: 50}} 
                     title="Cancelar" 
-                    onPress={handleEditar} />
+                    onPress={handleCadastrar} />
             </View>
 
         </View>
@@ -141,4 +141,4 @@ const EditarServico = ({ navigation }: any) => {
 };
 
 
-export default EditarServico;
+export default CadastrarServico;
