+++ package.json
@@ -1,7 +1,8 @@
 {
   "name": "mechapp",
   "version": "1.0.0",
-  "main": "index.ts",
+  "main": "expo-router/entry",
+
   "scripts": {
     "start": "expo start",
     "android": "expo start --android",
@@ -18,7 +19,11 @@
     "react": "19.0.0",
     "react-dom": "19.0.0",
     "react-native": "0.79.3",
-    "react-native-web": "^0.20.0"
+    "react-native-web": "^0.20.0",
+    "expo-router": "~5.0.7",
+    "react-native-screens": "~4.11.1",
+    "react-native-safe-area-context": "5.4.0",
+    "react-native-gesture-handler": "~2.24.0"
   },
   "devDependencies": {
     "@babel/core": "^7.25.2",
